senior1code/backend/requirements.txt
fastapi
uvicorn
pydantic
huggingface_hub>=0.22.0
python-dotenv